small fixes

+++ 00_scripts/slurm_code/03_gsnap_array.sh
@@ -68,4 +68,4 @@ else
         fi
 fi
 
-rm samples
+if [ -e samples ]; then rm samples ; fi
